feat: 온라인 대전모드 시작화면으로 이동 난이도와 게임 모드를 선택할 필요가 없기에 시작화면으로 이동하였습니다.
