+++ app/src/main/resources/org/se13/StartScreen.fxml
@@ -10,6 +10,7 @@
       fx:controller="org.se13.view.start.StartScreenController">
     <Text text="Tetris" styleClass="gameTitle"/>
     <Button fx:id="startButton" text="Start" styleClass="menuButton" onAction="#handleTetrisButtonAction"/>
+    <Button fx:id="onlineButton" text="Online" styleClass="menuButton" onAction="#handleOnlineButtonAction"/>
     <Button fx:id="settingsButton" text="SETTINGS" styleClass="menuButton" onAction="#handleSettingsButtonAction"/>
     <Button fx:id="scoreButton" text="SCORE" styleClass="menuButton" onAction="#handleScoreButtonAction"/>
     <Button fx:id="quitButton" text="QUIT" styleClass="menuButton" onAction="#handleQuitButtonAction"/>
